Refactor generateEmissionsTable to .ts
