+++ README.md
@@ -12,6 +12,6 @@ See https://carbon.fyi/learn
 The emission table needs periodic updating to track the most recent chain usage. Download .csv gas used data from https://etherscan.io/chart/gasused and hashrate data from https://etherscan.io/chart/hashrate
 To generate up-to-date emission data run:
 ```
-node src/data/generateEmissionsTable.js
+node src/data/generateEmissionsTable.ts
 ```
