+++ babel.config.cjs
@@ -3,5 +3,5 @@ module.exports = {
   presets: [
     ["@babel/preset-env", { targets: { node: "current" } }],
     "@babel/preset-typescript",
-  ],
+  ]
 };
